Update CartController and ProductService.

+++ src/main/java/com/ecommerce/ShoppingCart/Dto/AddToCartDto.java
@@ -6,13 +6,13 @@ import com.sun.istack.NotNull;
 public class AddToCartDto {
     private Integer id;
     private @NotNull Integer userId;
-    private @NotNull Long productId;
+    private @NotNull Integer productId;
     private @NotNull Integer quantity;
 
     public AddToCartDto(){
     }
 
-    public AddToCartDto(Integer id, @NotNull Integer userId, @NotNull Long productId, @NotNull Integer quantity) {
+    public AddToCartDto(Integer id, @NotNull Integer userId, @NotNull Integer productId, @NotNull Integer quantity) {
         this.id = id;
         this.userId = userId;
         this.productId = productId;
@@ -42,11 +42,11 @@ public class AddToCartDto {
         this.userId = userId;
     }
 
-    public Long getProductId() {
+    public Integer getProductId() {
         return productId;
     }
 
-    public void setProductId(Long productId) {
+    public void setProductId(Integer productId) {
         this.productId = productId;
     }
